refactor: extract BindingMixin from DuiCard and DuiKey
Extract duplicated bind/bind_range/bind_many/forward/_wrap_handler,
on, bind_event, and detach methods into a shared BindingMixin class.

Both DuiCard and DuiKey now inherit from BindingMixin, eliminating
mirror-copy implementations and reducing divergence risk.

No change to public API. All tests pass with coverage maintained.

Closes #212
